Added flash times and good runs for NuMI

+++ selection/include/muon2024/cuts_muon2024.h
@@ -29,7 +29,7 @@
  */
 namespace cuts::muon2024
 {
-    std::vector<float> flash_times;
+    std::vector<double> flash_times;
     static const bool initialized = [] {
         if constexpr(!BEAM_IS_NUMI)
             flash_times = {0.0,1.6};
